fix: lock deps/devDeps versions
It was taking too much time to set up a new project with `npm create
ssam@latest` because it had to check for latest versions for many
packages and was not using cached versions if any.

Now, all deps versions are locked and install time is much faster. I
just need to keep updating versions whenever there's a change. But users
can always update themselves, too.